switched to flatter structure
