+++ lite/benchmarks/benchmark_objdet.py
@@ -8,7 +8,7 @@ from time import time
 
 import requests
 from tqdm import tqdm
-from valor_lite.cv.detection import DataLoader, MetricType
+from valor_lite.object_detection import DataLoader, MetricType
 
 
 class AnnotationType(str, Enum):
