+++ lite/tests/object_detection/test_counts.py
@@ -1,7 +1,7 @@
-from valor_lite.cv.detection import DataLoader, Detection, MetricType
+from valor_lite.object_detection import DataLoader, Detection, MetricType
 
 
-def test_precision_metrics_first_class(
+def test_counts_metrics_first_class(
     basic_detections_first_class: list[Detection],
     basic_rotated_detections_first_class: list[Detection],
 ):
@@ -11,7 +11,6 @@ def test_precision_metrics_first_class(
     groundtruths
         datum uid1
             box 1 - label v1 - tp
-            box 3 - label v2 - fn missing prediction
         datum uid2
             box 2 - label v1 - fn missing prediction
 
@@ -19,8 +18,9 @@ def test_precision_metrics_first_class(
         datum uid1
             box 1 - label v1 - score 0.3 - tp
         datum uid2
-            box 2 - label v2 - score 0.98 - fp
+           none
     """
+
     for input_, method in [
         (basic_detections_first_class, DataLoader.add_bounding_boxes),
         (basic_rotated_detections_first_class, DataLoader.add_polygons),
@@ -42,12 +42,16 @@ def test_precision_metrics_first_class(
         assert evaluator.n_groundtruths == 2
         assert evaluator.n_predictions == 1
 
-        # test Precision
-        actual_metrics = [m for m in metrics[MetricType.Precision]]
+        # test Counts
+        actual_metrics = [m for m in metrics[MetricType.Counts]]
         expected_metrics = [
             {
-                "type": "Precision",
-                "value": 1.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 1,
+                    "fp": 0,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.1,
                     "score_threshold": 0.0,
@@ -55,8 +59,12 @@ def test_precision_metrics_first_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 1.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 1,
+                    "fp": 0,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.6,
                     "score_threshold": 0.0,
@@ -64,8 +72,12 @@ def test_precision_metrics_first_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 0,
+                    "fn": 2,
+                },
                 "parameters": {
                     "iou_threshold": 0.1,
                     "score_threshold": 0.5,
@@ -73,8 +85,12 @@ def test_precision_metrics_first_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 0,
+                    "fn": 2,
+                },
                 "parameters": {
                     "iou_threshold": 0.6,
                     "score_threshold": 0.5,
@@ -88,7 +104,7 @@ def test_precision_metrics_first_class(
             assert m in actual_metrics
 
 
-def test_precision_metrics_second_class(
+def test_counts_metrics_second_class(
     basic_detections_second_class: list[Detection],
     basic_rotated_detections_second_class: list[Detection],
 ):
@@ -106,6 +122,7 @@ def test_precision_metrics_second_class(
         datum uid2
             box 2 - label v2 - score 0.98 - fp
     """
+
     for input_, method in [
         (basic_detections_second_class, DataLoader.add_bounding_boxes),
         (basic_rotated_detections_second_class, DataLoader.add_polygons),
@@ -117,7 +134,6 @@ def test_precision_metrics_second_class(
         metrics = evaluator.evaluate(
             iou_thresholds=[0.1, 0.6],
             score_thresholds=[0.0, 0.5],
-            as_dict=True,
         )
 
         assert evaluator.ignored_prediction_labels == []
@@ -127,12 +143,16 @@ def test_precision_metrics_second_class(
         assert evaluator.n_groundtruths == 1
         assert evaluator.n_predictions == 1
 
-        # test Precision
-        actual_metrics = [m for m in metrics[MetricType.Precision]]
+        # test Counts
+        actual_metrics = [m.to_dict() for m in metrics[MetricType.Counts]]
         expected_metrics = [
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 1,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.1,
                     "score_threshold": 0.0,
@@ -140,8 +160,12 @@ def test_precision_metrics_second_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 1,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.6,
                     "score_threshold": 0.0,
@@ -149,8 +173,12 @@ def test_precision_metrics_second_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 1,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.1,
                     "score_threshold": 0.5,
@@ -158,8 +186,12 @@ def test_precision_metrics_second_class(
                 },
             },
             {
-                "type": "Precision",
-                "value": 0.0,
+                "type": "Counts",
+                "value": {
+                    "tp": 0,
+                    "fp": 1,
+                    "fn": 1,
+                },
                 "parameters": {
                     "iou_threshold": 0.6,
                     "score_threshold": 0.5,
@@ -173,7 +205,7 @@ def test_precision_metrics_second_class(
             assert m in actual_metrics
 
 
-def test_precision_false_negatives_single_datum_baseline(
+def test_counts_false_negatives_single_datum_baseline(
     false_negatives_single_datum_baseline_detections: list[Detection],
 ):
     """This is the baseline for the below test. In this case there are two predictions and
@@ -191,11 +223,15 @@ def test_precision_false_negatives_single_datum_baseline(
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -203,8 +239,12 @@ def test_precision_false_negatives_single_datum_baseline(
             },
         },
         {
-            "type": "Precision",
-            "value": 0.0,
+            "type": "Counts",
+            "value": {
+                "tp": 0,
+                "fp": 0,
+                "fn": 1,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.9,
@@ -218,7 +258,7 @@ def test_precision_false_negatives_single_datum_baseline(
         assert m in actual_metrics
 
 
-def test_precision_false_negatives_single_datum(
+def test_counts_false_negatives_single_datum(
     false_negatives_single_datum_detections: list[Detection],
 ):
     """Tests where high confidence false negative was not being penalized. The
@@ -235,11 +275,15 @@ def test_precision_false_negatives_single_datum(
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -253,7 +297,7 @@ def test_precision_false_negatives_single_datum(
         assert m in actual_metrics
 
 
-def test_precision_false_negatives_two_datums_one_empty_low_confidence_of_fp(
+def test_counts_false_negatives_two_datums_one_empty_low_confidence_of_fp(
     false_negatives_two_datums_one_empty_low_confidence_of_fp_detections: list[
         Detection
     ],
@@ -278,11 +322,15 @@ def test_precision_false_negatives_two_datums_one_empty_low_confidence_of_fp(
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -290,13 +338,13 @@ def test_precision_false_negatives_two_datums_one_empty_low_confidence_of_fp(
             },
         }
     ]
-    for m in actual_metrics:
-        assert m in expected_metrics
     for m in expected_metrics:
         assert m in actual_metrics
+    for m in actual_metrics:
+        assert m in expected_metrics
 
 
-def test_precision_false_negatives_two_datums_one_empty_high_confidence_of_fp(
+def test_counts_false_negatives_two_datums_one_empty_high_confidence_of_fp(
     false_negatives_two_datums_one_empty_high_confidence_of_fp_detections: list[
         Detection
     ],
@@ -320,11 +368,15 @@ def test_precision_false_negatives_two_datums_one_empty_high_confidence_of_fp(
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -332,13 +384,13 @@ def test_precision_false_negatives_two_datums_one_empty_high_confidence_of_fp(
             },
         }
     ]
-    for m in actual_metrics:
-        assert m in expected_metrics
     for m in expected_metrics:
         assert m in actual_metrics
+    for m in actual_metrics:
+        assert m in expected_metrics
 
 
-def test_precision_false_negatives_two_datums_one_only_with_different_class_low_confidence_of_fp(
+def test_counts_false_negatives_two_datums_one_only_with_different_class_low_confidence_of_fp(
     false_negatives_two_datums_one_only_with_different_class_low_confidence_of_fp_detections: list[
         Detection
     ],
@@ -362,11 +414,15 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_low_
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -374,8 +430,12 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_low_
             },
         },
         {
-            "type": "Precision",
-            "value": 0.0,
+            "type": "Counts",
+            "value": {
+                "tp": 0,
+                "fp": 0,
+                "fn": 1,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -383,13 +443,13 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_low_
             },
         },
     ]
-    for m in actual_metrics:
-        assert m in expected_metrics
     for m in expected_metrics:
         assert m in actual_metrics
+    for m in actual_metrics:
+        assert m in expected_metrics
 
 
-def test_precision_false_negatives_two_datums_one_only_with_different_class_high_confidence_of_fp(
+def test_counts_false_negatives_two_datums_one_only_with_different_class_high_confidence_of_fp(
     false_negatives_two_images_one_only_with_different_class_high_confidence_of_fp_detections: list[
         Detection
     ],
@@ -413,11 +473,15 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_high
         as_dict=True,
     )
 
-    actual_metrics = [m for m in metrics[MetricType.Precision]]
+    actual_metrics = [m for m in metrics[MetricType.Counts]]
     expected_metrics = [
         {
-            "type": "Precision",
-            "value": 0.5,
+            "type": "Counts",
+            "value": {
+                "tp": 1,
+                "fp": 1,
+                "fn": 0,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -425,8 +489,12 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_high
             },
         },
         {
-            "type": "Precision",
-            "value": 0.0,
+            "type": "Counts",
+            "value": {
+                "tp": 0,
+                "fp": 0,
+                "fn": 1,
+            },
             "parameters": {
                 "iou_threshold": 0.5,
                 "score_threshold": 0.0,
@@ -434,7 +502,108 @@ def test_precision_false_negatives_two_datums_one_only_with_different_class_high
             },
         },
     ]
-    for m in actual_metrics:
-        assert m in expected_metrics
     for m in expected_metrics:
         assert m in actual_metrics
+    for m in actual_metrics:
+        assert m in expected_metrics
+
+
+def test_counts_ranked_pair_ordering(
+    detection_ranked_pair_ordering: Detection,
+    detection_ranked_pair_ordering_with_bitmasks: Detection,
+    detection_ranked_pair_ordering_with_polygons: Detection,
+):
+
+    for input_, method in [
+        (detection_ranked_pair_ordering, DataLoader.add_bounding_boxes),
+        (
+            detection_ranked_pair_ordering_with_bitmasks,
+            DataLoader.add_bitmasks,
+        ),
+        (
+            detection_ranked_pair_ordering_with_polygons,
+            DataLoader.add_polygons,
+        ),
+    ]:
+        loader = DataLoader()
+        method(loader, detections=[input_])
+        evaluator = loader.finalize()
+
+        assert evaluator.metadata == {
+            "ignored_prediction_labels": [
+                "label4",
+            ],
+            "missing_prediction_labels": [],
+            "n_datums": 1,
+            "n_groundtruths": 3,
+            "n_labels": 4,
+            "n_predictions": 4,
+        }
+
+        metrics = evaluator.evaluate(
+            iou_thresholds=[0.5, 0.75],
+            score_thresholds=[0.0],
+            as_dict=True,
+        )
+
+        actual_metrics = [m for m in metrics[MetricType.Counts]]
+        expected_metrics = [
+            {
+                "type": "Counts",
+                "value": {"tp": 1, "fp": 0, "fn": 0},
+                "parameters": {
+                    "iou_threshold": 0.5,
+                    "score_threshold": 0.0,
+                    "label": "label1",
+                },
+            },
+            {
+                "type": "Counts",
+                "value": {"tp": 1, "fp": 0, "fn": 0},
+                "parameters": {
+                    "iou_threshold": 0.75,
+                    "score_threshold": 0.0,
+                    "label": "label1",
+                },
+            },
+            {
+                "type": "Counts",
+                "value": {"tp": 1, "fp": 0, "fn": 0},
+                "parameters": {
+                    "iou_threshold": 0.5,
+                    "score_threshold": 0.0,
+                    "label": "label2",
+                },
+            },
+            {
+                "type": "Counts",
+                "value": {"tp": 1, "fp": 0, "fn": 0},
+                "parameters": {
+                    "iou_threshold": 0.75,
+                    "score_threshold": 0.0,
+                    "label": "label2",
+                },
+            },
+            {
+                "type": "Counts",
+                "value": {"tp": 0, "fp": 1, "fn": 1},
+                "parameters": {
+                    "iou_threshold": 0.5,
+                    "score_threshold": 0.0,
+                    "label": "label3",
+                },
+            },
+            {
+                "type": "Counts",
+                "value": {"tp": 0, "fp": 1, "fn": 1},
+                "parameters": {
+                    "iou_threshold": 0.75,
+                    "score_threshold": 0.0,
+                    "label": "label3",
+                },
+            },
+        ]
+        for m in actual_metrics:
+            assert m in expected_metrics
+        for m in expected_metrics:
+            assert m in actual_metrics
