+++ lite/tests/semantic_segmentation/test_dataloader.py
@@ -1,5 +1,5 @@
 import pytest
-from valor_lite.cv.segmentation import DataLoader
+from valor_lite.semantic_segmentation import DataLoader
 
 
 def test_no_data():
